+++ tests/testthat/test-plots-plate.R
@@ -26,10 +26,10 @@ get_test_plate <- function() {
 
 get_real_plate <- function() {
   plate_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv",
-    package = "PvSTATEM"
+    package = "SerolyzeR"
   )
   layout_filepath <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx",
-    package = "PvSTATEM"
+    package = "SerolyzeR"
   )
 
   read_luminex_data(plate_filepath, layout_filepath)