+++ tests/testthat/test-plots-standard_curve.R
@@ -31,7 +31,7 @@ get_test_list_of_plates <- function() {
 }
 
 get_list_of_plates <- function() {
-  dir <- system.file("extdata", "multiplate_reallife_reduced", package = "PvSTATEM", mustWork = TRUE)
+  dir <- system.file("extdata", "multiplate_reallife_reduced", package = "SerolyzeR", mustWork = TRUE)
 
   output_dir <- tempdir(check = TRUE)
   dir.create(output_dir)